refactor: group buttons
